backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1
aioredis==2.0.1

# Data processing and numerical computing
numpy==1.26.2
pandas==2.1.4
numba==0.58.1
scipy==1.11.4

# Financial data and indicators
alpaca-trade-api==3.1.1
polygon-api-client==1.13.0
yfinance==0.2.22
ta==0.10.2

# Machine Learning
scikit-learn==1.3.2
torch==2.1.2
transformers==4.36.2

# High-performance JSON
orjson==3.9.10

# WebSocket and networking
aiohttp==3.9.1
httpx==0.25.2

# Monitoring and observability
prometheus-client==0.19.0
structlog==23.2.0

# Caching and serialization
msgpack==1.0.7
lz4==4.3.2

# Configuration and utilities
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
click==8.1.7

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Time and date handling
pytz==2023.3
arrow==1.3.0